Extract `ConfigReaderException` pretty printing logic
